Extract a test implementation of IDynamicFileInfoProvider into test utilties
