+++ src/Workspaces/CoreTest/WorkspaceTests/DynamicFileInfoProviderMefTests.cs
@@ -3,10 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Composition;
 using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Test.Utilities;
@@ -45,37 +42,12 @@ namespace Microsoft.CodeAnalysis.UnitTests
             });
         }
 
-        #region Helpers
         internal static Lazy<IDynamicFileInfoProvider, FileExtensionsMetadata> GetDynamicFileInfoProvider()
         {
-            var catalog = ExportProviderCache.CreateTypeCatalog(new Type[] { typeof(TestDynamicFileInfoProvider) });
+            var catalog = ExportProviderCache.CreateTypeCatalog(new Type[] { typeof(TestDynamicFileInfoProviderThatProducesNoFiles) });
             var factory = ExportProviderCache.GetOrCreateExportProviderFactory(catalog);
 
             return factory.CreateExportProvider().GetExport<IDynamicFileInfoProvider, FileExtensionsMetadata>();
         }
-
-        [ExportDynamicFileInfoProvider("cshtml", "vbhtml")]
-        [Shared]
-        [PartNotDiscoverable]
-        internal class TestDynamicFileInfoProvider : IDynamicFileInfoProvider
-        {
-            [ImportingConstructor]
-            [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-            public TestDynamicFileInfoProvider()
-            {
-            }
-
-            public event EventHandler<string> Updated;
-
-            public Task<DynamicFileInfo> GetDynamicFileInfoAsync(ProjectId projectId, string projectFilePath, string filePath, CancellationToken cancellationToken)
-                => Task.FromResult<DynamicFileInfo>(null);
-
-            public Task RemoveDynamicFileInfoAsync(ProjectId projectId, string projectFilePath, string filePath, CancellationToken cancellationToken)
-                => Task.CompletedTask;
-
-            private void OnUpdate()
-                => Updated?.Invoke(this, "test");
-        }
-        #endregion
     }
 }
